FEAT: added comment notfication and deduped code by moving into common file
